Add files via upload
This script was very very useful for cleaning and organizing excels tailored to the customers needs. The most useful part is the Keyword extraction. The data was exported out of Scooper's platform in a raw form that didn't show what relevant keywords were inside the strings in the Post Content column. I really wanted these words to pop out to the customer, so they can extract meaningful insights from it as quickly as possible.